added a default permissive policy

+++ src/App/Http/Controllers/Company/Destroy.php
@@ -2,13 +2,18 @@
 
 namespace LaravelEnso\Companies\App\Http\Controllers\Company;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Routing\Controller;
 use LaravelEnso\Companies\App\Models\Company;
 
 class Destroy extends Controller
 {
+    use AuthorizesRequests;
+
     public function __invoke(Company $company)
     {
+        $this->authorize('destroy', $company);
+
         $company->delete();
 
         return [
